Added Inventory Excercise
I added an Inventory exercise where an existing inventory is updated according to a new inventory received. If the current inventory does not have an item from the new inventory, we add that item with its quantity. If it has already that product then we just simply update the quantity. The printed result is printed alphabetically